Remove race condition
When installing apps, the global config is changed to allow installing
applications, occ is used to install the application, and then the
config is changed again to disable installing applications.

This means that during the use of the install task, we have a race
condition that allows to install from the GUI, what we do not want to
allow.

Instead of tuning the configuration, only enable apps management if we
are running from the context of a task.